+++ src/link/LinkBattle.lua
@@ -47,6 +47,7 @@ end
 local BUILTIN_RULESETS = {
   gen1_faithful = require("src.battle.rulesets.gen1_faithful"),
   modern_clean = require("src.battle.rulesets.modern_clean"),
+  modern = require("src.battle.rulesets.modern"),
 }
 
 local function rulesetFor(game, id)